requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
duckdb==0.10.0
pandas==2.1.4
pyarrow==14.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.26.0
